LayoutTests/crypto/subtle/aes-cbc-import-key-unwrap-jwk-rsa-key-public-expected.txt

Test unwrapping a JWK RSA public key with AES-CBC using an imported key

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS unwrappedKey.kty is jwkKey.kty
PASS unwrappedKey.alg is jwkKey.alg
PASS unwrappedKey.key_ops is jwkKey.key_ops
PASS unwrappedKey.ext is jwkKey.ext
PASS unwrappedKey.n is jwkKey.n
PASS unwrappedKey.e is jwkKey.e
PASS successfullyParsed is true

TEST COMPLETE
